Added ACTDR6 data and examples. Also lots of optimizations on setup_cocoa - data won't be downloaded every time you call the shell script

+++ Cocoa/installation_scripts/setup_core_packages.sh
@@ -57,43 +57,37 @@ if [ -z "${IGNORE_CORE_INSTALLATION}" ]; then
 
     local FILE="${1:?}.${2:?}"
 
-    # ----------------------------------------------------------------------------
-    # In case this script runs twice (after being killed by CTRL-D)
-    rm -rf "${CCIL:?}/${1:?}"
-    rm -f  "${CCIL:?}/${FILE:?}"
-    # ----------------------------------------------------------------------------
-
-    wget "${3:?}" --retry-connrefused --waitretry=1 --tries=3 --read-timeout=20 \
-      --timeout=15 --waitretry=0 --show-progress --progress=bar:force \
-      >${OUT1:?} 2>${OUT2:?} || { error "${EC24:?}"; return 1; }
-
-    if [ "${2:?}" == "tar.gz" ]; then
-      
-      tar zxvf "${FILE}" \
-        >${OUT1:?} 2>${OUT2:?} || { error "${EC25:?} (gz)"; return 1; }
-    
-    elif [ "${2:?}" == "tar.xz" ]; then
-    
-      tar xf "${FILE}" \
-        >${OUT1:?} 2>${OUT2:?} || { error "${EC25:?} (xz)"; return 1; }
-    
-    else
-
-      error "UNKNOWN FILE EXTENSION"; return 1;
-    
-    fi
-
-    if [ "${1:?}/" != "${4:?}" ] && [ "${1:?}" != "${4:?}" ] \
-      && [ "${1:?}" != "${4:?}/" ] ; then
-
-      # In case this script runs twice (after being killed by CTRL-D)
+    # In case this script runs twice -------------------------------------------
+    if [ -n "${OVERWRITE_EXISTING_CORE_PACKAGES}" ]; then
       rm -rf "${CCIL:?}/${4:?}"
-      
-      mv "${1:?}/" "${4:?}" 2>${OUT2:?} || { error "MV FOLDER"; return 1; }
-    
+      rm -rf "${CCIL:?}/${1:?}"
+      if [ -n "${REDOWNLOAD_EXISTING_CORE_PACKAGES}" ]; then
+        rm -f  "${CCIL:?}/${FILE:?}"
+      fi 
     fi
 
-    rm -f  "${CCIL:?}/${FILE:?}"
+    if [[ ! -d "${CCIL:?}/${4:?}" && ! -d "${CCIL:?}/${1:?}" ]]; then
+
+      if [[ ! -e "${CCIL:?}/${FILE:?}" ]]; then
+        wget "${3:?}" --retry-connrefused --waitretry=1 --tries=3 --read-timeout=20 \
+          --timeout=15 --waitretry=0 --show-progress --progress=bar:force \
+          >${OUT1:?} 2>${OUT2:?} || { error "${EC24:?}"; return 1; }
+      fi
+
+      if [ "${2:?}" == "tar.gz" ]; then
+        tar zxvf "${FILE}" >${OUT1:?} 2>${OUT2:?} || { error "${EC25:?} (gz)"; return 1; }
+      elif [ "${2:?}" == "tar.xz" ]; then
+        tar xf "${FILE}" >${OUT1:?} 2>${OUT2:?} || { error "${EC25:?} (xz)"; return 1; }
+      else
+        error "UNKNOWN FILE EXTENSION"; return 1;
+      fi
+
+      if [[ "${1:?}/" != "${4:?}" && "${1:?}" != "${4:?}" && "${1:?}" != "${4:?}/" ]]; then
+        # In case this script runs twice (after being killed by CTRL-D)
+        rm -rf "${CCIL:?}/${4:?}"
+        mv "${1:?}/" "${4:?}" 2>${OUT2:?} || { error "MV FOLDER"; return 1; }
+      fi
+    fi
 
     cdfolder "${ROOTDIR}" || return 1;
   }
@@ -105,19 +99,22 @@ if [ -z "${IGNORE_CORE_INSTALLATION}" ]; then
 
     cdfolder "${CCIL:?}" || return 1;
 
-    # ----------------------------------------------------------------------------
-    # In case this script runs twice (after being killed by CTRL-D)
-    rm -f  "${CCIL:?}/${5:?}"
-    # ----------------------------------------------------------------------------
+    # In case this script runs twice -------------------------------------------
+    if [ -n "${OVERWRITE_EXISTING_CORE_PACKAGES}" ]; then
+    
+      rm -f  "${CCIL:?}/${5:?}"
+    
+    fi
 
-    # Why this compress? In the old Cocoa we saved the libraries in 
-    # the github repo using git lfs. So this way preserves the old scripts
-    # we set "-k 1" to be the minimum compression
-    tar -cf - "${4:?}" | xz -k -1 --threads=$MNT -c - > "${5}" \
-      2>${OUT2:?} || { error "TAR (COMPRESS)"; return 1; }
+    if [ ! -e "${CCIL:?}/${5:?}" ]; then
+    
+      # Why this compress? In the old Cocoa we saved the libraries in 
+      # the github repo using git lfs. So this way preserves the old scripts
+      # we set "-k 1" to be the minimum compression
+      tar -cf - "${4:?}" | xz -k -1 --threads=$MNT -c - > "${5}" \
+        2>${OUT2:?} || { error "TAR (COMPRESS)"; return 1; }
 
-    rm -rf "${CCIL:?}/${1:?}"
-    rm -rf "${CCIL:?}/${4:?}"
+    fi
 
     cdfolder "${ROOTDIR}" || return 1;
   }
@@ -138,23 +135,29 @@ if [ -z "${IGNORE_CORE_INSTALLATION}" ]; then
 
     cdfolder "${CCIL:?}" || return 1;
 
-    # ---------------------------------------------------------------------------
-    # In case this script runs twice --------------------------------------------
-    rm -rf "${CCIL:?}/${1:?}"
-    # ---------------------------------------------------------------------------
+    # In case this script runs twice -------------------------------------------
+    if [ -n "${OVERWRITE_EXISTING_CORE_PACKAGES}" ]; then
+    
+      rm -rf "${CCIL:?}/${1:?}"
+    
+    fi
 
-    "${CURL:?}" -fsS "${3:?}" \
-      >${OUT1:?} 2>${OUT2:?} || { error "${EC27:?} (URL=${3:?})"; return 1; }
+    if [ ! -d "${CCIL:?}/${1:?}" ]; then
+    
+      "${CURL:?}" -fsS "${3:?}" \
+        >${OUT1:?} 2>${OUT2:?} || { error "${EC27:?} (URL=${3:?})"; return 1; }
 
-    "${GIT:?}" clone "${3:?}" "${1:?}" \
-      >${OUT1:?} 2>${OUT2:?} || { error "GIT CLONE"; return 1; }
+      "${GIT:?}" clone "${3:?}" "${1:?}" \
+        >${OUT1:?} 2>${OUT2:?} || { error "GIT CLONE"; return 1; }
+      
+      cdfolder "${CCIL:?}/${1:?}" || return 1;
+      
+      "${GIT:?}" checkout "${2:?}" \
+        >${OUT1:?} 2>${OUT2:?} || { error "GIT CHECKOUT"; return 1; }
+      
+      rm -rf "${CCIL:?}/${1:?}/.git/"
     
-    cdfolder "${CCIL:?}/${1:?}" || return 1;
-    
-    "${GIT:?}" checkout "${2:?}" \
-      >${OUT1:?} 2>${OUT2:?} || { error "GIT CHECKOUT"; return 1; }
-    
-    rm -rf "${CCIL:?}/${1:?}/.git/"
+    fi
 
     cdfolder "${ROOTDIR}" || return 1;
   }
@@ -164,30 +167,34 @@ if [ -z "${IGNORE_CORE_INSTALLATION}" ]; then
     
     local CCIL="${ROOTDIR}/../cocoa_installation_libraries"
 
-    # In case this script runs twice (after being killed by CTRL-D)
-    rm -f  "${CCIL:?}/${3:?}"
+    # In case this script runs twice -------------------------------------------
+    if [ -n "${OVERWRITE_EXISTING_CORE_PACKAGES}" ]; then
 
-    cdfolder "${CCIL:?}" || return 1;
+      rm -f  "${CCIL:?}/${3:?}"
 
-    if [ "${1:?}/" != "${2:?}" ] && [ "${1:?}" != "${2:?}" ] \
-      && [ "${1:?}" != "${2:?}/" ]; then
-
-      # In case this script runs twice (after being killed by CTRL-D)
-      rm -rf "${CCIL:?}/${2:?}"
-
-      mv "${1:?}/" "${2:?}" 2>${OUT2:?} || { error "MV FOLDER"; return 1; }
-    
     fi
 
-    # Why this compress? In the old Cocoa we saved the libraries in 
-    # the github repo using git lfs. So this way preserves the old scripts
-    # we set "-k 1" to be the minimum compression
-    tar -cf - "${2:?}" | xz -k -1 --threads=$MNT -c - > "${3}" \
-      2>${OUT2:?} || { error "TAR (COMPRESS)"; return 1; }
-
-    rm -rf "${CCIL:?}/${1:?}"
+    if [ ! -d "${CCIL:?}/${3:?}" ]; then
     
-    rm -rf "${CCIL:?}/${2:?}"
+      cdfolder "${CCIL:?}" || return 1;
+
+      if [ "${1:?}/" != "${2:?}" ] && [ "${1:?}" != "${2:?}" ] \
+        && [ "${1:?}" != "${2:?}/" ]; then
+
+        # In case this script runs twice (after being killed by CTRL-D)
+        rm -rf "${CCIL:?}/${2:?}"
+
+        mv "${1:?}/" "${2:?}" 2>${OUT2:?} || { error "MV FOLDER"; return 1; }
+      
+      fi
+
+      # Why this compress? In the old Cocoa we saved the libraries in 
+      # the github repo using git lfs. So this way preserves the old scripts
+      # we set "-k 1" to be the minimum compression
+      tar -cf - "${2:?}" | xz -k -1 --threads=$MNT -c - > "${3}" \
+        2>${OUT2:?} || { error "TAR (COMPRESS)"; return 1; }
+
+    fi
 
     cdfolder "${ROOTDIR}" || return 1;
   }
@@ -212,6 +219,7 @@ if [ -z "${IGNORE_CORE_INSTALLATION}" ]; then
   # ----------------------------------------------------------------------------
 
   if [ -z "${IGNORE_XZ_INSTALLATION}" ]; then
+  
     ptop "GETTING AND COMPILING XZ LIBRARY (CORE LIBS)" || return 1;
 
     cdfolder "${CCIL:?}" || return 1;
@@ -235,6 +243,7 @@ if [ -z "${IGNORE_CORE_INSTALLATION}" ]; then
     cdfolder "${ROOTDIR}" || return 1;
 
     pbottom "GETTING AND COMPILING XZ LIBRARY (CORE LIBS)" || return 1;
+  
   fi
 
   # ----------------------------------------------------------------------------
@@ -674,31 +683,38 @@ if [ -z "${IGNORE_CORE_INSTALLATION}" ]; then
 
     XZF="carma.xz"
 
-    gitact1 "${FOLDER:?}" "${VER:?}" "${URL:?}" || return 1;
-
-    # -------------------------------------------------------------------------
-    # In case this script runs twice (after sudden break w/ CTRL-C)
-    rm -rf "${PACKDIR:?}"
-    rm -rf "${CCIL:?}/include"
-
-    # -------------------------------------------------------------------------
-    # move/rename include file and carma.h folder
-    # -------------------------------------------------------------------------
-    mv "${CCIL:?}/${FOLDER:?}/include" "${CCIL:?}" >${OUT1:?} 2>${OUT2:?} || 
-      { error "MV CARMA INCLUDE FOLDER"; return 1; }
-
-    mv "${CCIL:?}/include" "${PACKDIR:?}" 2>${OUT2:?} || 
-      { error "RENAME CARMA INCLUDE FOLDER"; return 1; }
-
-    mv "${PACKDIR:?}/carma" "${PACKDIR:?}/carma.h" \
-      2>${OUT2:?} || { error "RENANE CARMA HEADER"; return 1; }
-
-    rm -rf "${CCIL:?}/${FOLDER:?}"
-    # -------------------------------------------------------------------------
+    # In case this script runs twice -------------------------------------------
+    if [ -n "${OVERWRITE_EXISTING_CORE_PACKAGES}" ]; then
+      
+      rm -rf "${CCIL:?}/${FOLDER:?}"
+      rm -rf "${PACKDIR:?}"
+      rm -rf "${CCIL:?}/include"
     
-    gitact2 "${CNAME:?}" "${CNAME:?}" "${XZF:?}" || return 1; 
+    fi
 
-    unset -v CNAME PACKDIR URL FOLDER VER XZF
+    if [ ! -d "${CCIL:?}/${FOLDER:?}" ]; then
+      
+      gitact1 "${FOLDER:?}" "${VER:?}" "${URL:?}" || return 1;
+
+      # --------------------------------------------------------------------------
+      # move/rename include file and carma.h folder
+      # --------------------------------------------------------------------------
+      mv "${CCIL:?}/${FOLDER:?}/include" "${CCIL:?}" >${OUT1:?} 2>${OUT2:?} || 
+        { error "MV CARMA INCLUDE FOLDER"; return 1; }
+
+      mv "${CCIL:?}/include" "${PACKDIR:?}" 2>${OUT2:?} || 
+        { error "RENAME CARMA INCLUDE FOLDER"; return 1; }
+
+      mv "${PACKDIR:?}/carma" "${PACKDIR:?}/carma.h" \
+        2>${OUT2:?} || { error "RENANE CARMA HEADER"; return 1; }
+
+      # --------------------------------------------------------------------------
+      
+      gitact2 "${CNAME:?}" "${CNAME:?}" "${XZF:?}" || return 1; 
+
+      unset -v CNAME PACKDIR URL FOLDER VER XZF
+
+    fi
 
     pbottom "GETTING CARMA LIBRARY DONE (CORE LIBS)" || return 1;
