+++ Cocoa/installation_scripts/unxv_lipop.sh
@@ -12,7 +12,7 @@ if [ -z "${IGNORE_LIPOP_CMB_DATA}" ]; then
   ( source "${ROOTDIR:?}/installation_scripts/flags_check.sh" ) || return 1;
     
   unset_env_vars () { 
-    unset -v EDATAF URL LPDVS FOLDER PRINTNAME
+    unset -v EDATAF URL LPDVS FOLDER PRINTNAME TMP
     cdroot || return 1;
   }
 
@@ -38,13 +38,7 @@ if [ -z "${IGNORE_LIPOP_CMB_DATA}" ]; then
     cd "${1:?}" 2>"/dev/null" || { error "CD FOLDER: ${1}"; return 1; }
   }
 
-  #-----------------------------------------------------------------------------
-  #-----------------------------------------------------------------------------
-  #-----------------------------------------------------------------------------
-
   unset_env_vars || return 1
-
-  #-----------------------------------------------------------------------------
 
   # E = EXTERNAL, DATA, F=FODLER
   EDATAF="${ROOTDIR:?}/external_modules/data"
@@ -60,62 +54,42 @@ if [ -z "${IGNORE_LIPOP_CMB_DATA}" ]; then
                     "planck_2020_hillipop_TTTEEE_v${LPDVS:?}"
                     "planck_2020_lollipop")
 
-  # Name to be printed on this shell script messages
-  PRINTNAME="LIPOP"
-
-  # ---------------------------------------------------------------------------
-
-  ptop "SETUP/UNXV ${PRINTNAME:?} DATA" || return 1
-  
-  # ----------------------------------------------------------------------------
-  # note: in case script run >1x w/ previous run stoped prematurely b/c error
-  
-  rm -rf "${EDATAF:?}/planck/planck_2020"
-  
-  rm -rf "${EDATAF:?}/planck/hillipop"
-  
-  rm -rf "${EDATAF:?}/planck/lollipop"
-  
-  for (( i=0; i<${#FILE[@]}; i++ ));
-  do
-    rm -f  "${EDATAF:?}/planck/${FILE[$i]:?}.tar.gz"
-  done  
-
-  # ---------------------------------------------------------------------------
-  
-  cdfolder "${EDATAF:?}/planck"
-
-  for (( i=0; i<${#FILE[@]}; i++ ));
-  do
-
-    "${WGET:?}" "${URL}/${FILE[$i]:?}.tar.gz" -q --show-progress \
-      --progress=bar:force:noscroll || { error "${EC24:?}"; return 1; }
-
-    tar -xvzf "${FILE[$i]:?}.tar.gz" \
-      >${OUT1:?} 2>${OUT2:?} || { error "${EC25:?} (${FILE[$i]:?})"; return 1; }
-  
-  done
-
-  # ---------------------------------------------------------------------------
-  # note: by default, LIPOP saves the likelihoods on planck_2020
-  
-  mv "${EDATAF:?}/planck/planck_2020/hillipop" "${EDATAF:?}/planck"
-  
-  mv "${EDATAF:?}/planck/planck_2020/lollipop" "${EDATAF:?}/planck"
-
-  # ----------------------------------------------------------------------------
-  # clean tar.gz files 
-  
-  rm -rf "${EDATAF:?}/planck/planck_2020"
-
-  for (( i=0; i<${#FILE[@]}; i++ ));
-  do
-    rm -f  "${EDATAF:?}/planck/${FILE[$i]:?}.tar.gz"
-  done
-  
-  # ---------------------------------------------------------------------------
+  ptop "SETUP/UNXV LIPOP DATA" || return 1
     
-  pbottom "SETUP/UNXV ${PRINTNAME:?} DATA" || return 1
+  if [ -n "${OVERWRITE_EXISTING_LIPOP_CMB_DATA}" ]; then
+    rm -rf "${EDATAF:?}/planck/planck_2020"
+    rm -rf "${EDATAF:?}/planck/hillipop"  
+    rm -rf "${EDATAF:?}/planck/lollipop"
+    if [ -n "${REDOWNLOAD_EXISTING_LIPOP_CMB_DATA}" ]; then
+      for (( i=0; i<${#FILE[@]}; i++ ));
+      do
+        rm -f  "${EDATAF:?}/planck/${FILE[$i]:?}.tar.gz"
+      done 
+    fi
+  fi 
+  
+  if [[ ! -d "${EDATAF:?}/planck/hillipop" ||  
+        ! -d "${EDATAF:?}/planck/lollipop" ]]; then
+    cdfolder "${EDATAF:?}/planck"
+
+    for (( i=0; i<${#FILE[@]}; i++ ));
+    do
+      if [ ! -e "${EDATAF:?}/planck/${FILE[$i]:?}.tar.gz" ]; then
+        "${WGET:?}" "${URL}/${FILE[$i]:?}.tar.gz" -q --show-progress \
+          --progress=bar:force:noscroll || { error "${EC24:?}"; return 1; }
+      fi
+
+      tar -xvzf "${FILE[$i]:?}.tar.gz" \
+        >${OUT1:?} 2>${OUT2:?} || { error "${EC25:?} (${FILE[$i]:?})"; return 1; }
+    done
+
+    # note: by default, LIPOP saves the likelihoods on planck_2020
+    mv "${EDATAF:?}/planck/planck_2020/hillipop" "${EDATAF:?}/planck"
+    mv "${EDATAF:?}/planck/planck_2020/lollipop" "${EDATAF:?}/planck"
+    rm -rf "${EDATAF:?}/planck/planck_2020"
+  fi
+    
+  pbottom "SETUP/UNXV LIPOP DATA" || return 1
 
   unset_all || return 1;
